+++ plugin/v2/plugin_linux.go
@@ -7,8 +7,8 @@ import (
 	"strings"
 
 	"github.com/docker/docker/api/types"
-	"github.com/docker/docker/oci"
 	"github.com/docker/docker/pkg/system"
+	"github.com/docker/docker/runtime/oci"
 	"github.com/opencontainers/runtime-spec/specs-go"
 	"github.com/pkg/errors"
 )
