+++ runtime/oci/devices_linux.go
@@ -1,4 +1,4 @@
-package oci // import "github.com/docker/docker/oci"
+package oci // import "github.com/docker/docker/runtime/oci"
 
 import (
 	"fmt"
