+++ runtime/oci/devices_unsupported.go
@@ -1,6 +1,6 @@
 // +build !linux
 
-package oci // import "github.com/docker/docker/oci"
+package oci // import "github.com/docker/docker/runtime/oci"
 
 import (
 	"errors"
